chore: fix web tests
